requirements.txt
asgiref==3.4.1
backoff==1.10.0
boto3==1.18.32
botocore==1.21.32
Deprecated==1.2.12
googleapis-common-protos==1.53.0
grpcio==1.39.0
jmespath==0.10.0
MapProxy==1.13.2
numpy==1.21.5
opentelemetry-api==1.7.1
opentelemetry-exporter-otlp==1.7.1
opentelemetry-exporter-otlp-proto-grpc==1.7.1
opentelemetry-distro==0.26b1
opentelemetry-instrumentation==0.26b1
opentelemetry-instrumentation-boto==0.26b1
opentelemetry-instrumentation-botocore==0.26b1
opentelemetry-instrumentation-dbapi==0.26b1
opentelemetry-instrumentation-sqlite3==0.26b1
opentelemetry-instrumentation-wsgi==0.26b1
# opentelemetry-instrumentation-logging==0.26b1
opentelemetry-proto==1.7.1
opentelemetry-sdk==1.7.1
opentelemetry-semantic-conventions==0.26b1
Pillow==8.3.1
protobuf==3.17.3
pyproj==3.2.1
python-dateutil==2.8.2
PyYAML==5.4.1
s3transfer==0.5.0
Shapely==1.7.1
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.6
uWSGI==2.0.20
wrapt==1.12.1
wsgicors==0.7.0
